refactor: Reflected schemas refacto
