+++ src/app/api/deps.py
@@ -10,7 +10,8 @@ from pydantic import ValidationError
 
 import app.config as cfg
 from app.api import crud
-from app.api.schemas import AccessRead, AccessType, DeviceOut, TokenPayload, UserRead
+from app.api.schemas import AccessRead, DeviceOut, TokenPayload, UserRead
+from app.db.models import AccessType
 from app.db import accesses, devices, users
 
 # Scope definition
