+++ src/app/api/routes/installations.py
@@ -13,7 +13,8 @@ from app.api import crud
 from app.api.crud.authorizations import check_group_read, check_group_update, is_admin_access
 from app.api.crud.groups import get_entity_group_id
 from app.api.deps import get_current_access
-from app.api.schemas import AccessType, InstallationIn, InstallationOut
+from app.api.schemas import InstallationIn, InstallationOut
+from app.db.models import AccessType
 from app.db import get_session, installations, models
 
 router = APIRouter()
